requirements.txt
scikit-learn-1.2.2
customtkinter-5.1.2
seaborn-0.12.2